add checklogin

+++ BNHS/staff/search_item.php
@@ -17,6 +17,19 @@ if (isset($_GET['delete'])) {
     $err = "Try Again Later";
   }
 }
+// Check if a search query is submitted
+$searchResults = [];
+if (isset($_GET['query'])) {
+  $search = $mysqli->real_escape_string($_GET['query']); // Sanitize input to prevent SQL injection
+  $sql = "SELECT * FROM bnhs_staff WHERE staff_name LIKE '%$search%' OR staff_id LIKE '%$search%'";
+  $result = $mysqli->query($sql);
+
+  if ($result) {
+    while ($row = $result->fetch_object()) {
+      $searchResults[] = $row; // Store results in an array
+    }
+  }
+}
 require_once('partials/_head.php');
 ?>
 
@@ -32,8 +45,9 @@ require_once('partials/_head.php');
     require_once('partials/_topnav.php');
     ?>
     <!-- Header -->
-    <div style="background-image: url(assets/img/theme/bnhsfront.jpg); background-size: cover;" class="header  pb-8 pt-5 pt-md-8">
-    <span class="mask bg-gradient-dark opacity-8"></span>
+    <div style="background-image: url(assets/img/theme/bnhsfront.jpg); background-size: cover;"
+      class="header  pb-8 pt-5 pt-md-8">
+      <span class="mask bg-gradient-dark opacity-8"></span>
       <div class="container-fluid">
         <div class="header-body">
         </div>
@@ -46,10 +60,15 @@ require_once('partials/_head.php');
         <div class="col">
           <div class="card shadow">
             <div class="card-header border-0">
-              <a href="add_staff.php" class="btn btn-outline-success">
-                <i class="fas fa-user-plus"></i>
-                Add New Staff
-              </a>
+              <form class="form-inline" method="GET" action="search_staff.php"
+                style="float: left; margin-top: 20px; margin-bottom: 20px;">
+                <input class="form-control mr-sm-2" style="width: 500px;" type="search" name="query"
+                  placeholder="Search Item" aria-label="Search">
+                <button class="btn btn-outline-success my-2 my-sm-0" type="submit">
+                  <i class="fas fa-search"></i>
+                  Search
+                </button>
+              </form>
             </div>
             <div class="table-responsive">
               <table class="table align-items-center table-flush">
@@ -63,33 +82,37 @@ require_once('partials/_head.php');
                   </tr>
                 </thead>
                 <tbody>
-                  <?php
-                  $ret = "SELECT * FROM  bnhs_staff  ORDER BY `bnhs_staff`.`created_at` DESC ";
-                  $stmt = $mysqli->prepare($ret);
-                  $stmt->execute();
-                  $res = $stmt->get_result();
-                  while ($cust = $res->fetch_object()) {
-                  ?>
-                    <tr>
-                      <td><?php echo $cust->staff_id; ?></td>
-                      <td><?php echo $cust->staff_name; ?></td>
-                      <td><?php echo $cust->staff_phoneno; ?></td>
-                      <td><?php echo $cust->staff_email; ?></td>
-                      <td>
-                        <a href="user_management.php?delete=<?php echo $cust->staff_id; ?>">
+                  <?php if (!empty($searchResults)) { ?>
+                    <?php foreach ($searchResults as $cust) { ?>
+                      <tr>
+                        <td><?php echo $cust->staff_id; ?></td>
+                        <td><?php echo $cust->staff_name; ?></td>
+                        <td><?php echo $cust->staff_phoneno; ?></td>
+                        <td><?php echo $cust->staff_email; ?></td>
+                        <td>
+                          <a href="view_item.php?id=<?php echo $cust->staff_id; ?>">
+                            <button class="btn btn-sm btn-info">
+                              <i class="fas fa-eye"></i> View
+                            </button>
+                          </a>
+                          <a href="user_management.php?delete=<?php echo $cust->staff_id; ?>">
                           <button class="btn btn-sm btn-danger">
                             <i class="fas fa-trash"></i>
                             Delete
                           </button>
                         </a>
-
                         <a href="update_staff.php?update=<?php echo $cust->staff_id; ?>">
                           <button class="btn btn-sm btn-primary">
                             <i class="fas fa-user-edit"></i>
                             Update
                           </button>
                         </a>
-                      </td>
+                        </td>
+                      </tr>
+                    <?php } ?>
+                  <?php } else { ?>
+                    <tr>
+                      <td colspan="5" class="text-center">No items found matching your search.</td>
                     </tr>
                   <?php } ?>
                 </tbody>
@@ -98,8 +121,8 @@ require_once('partials/_head.php');
           </div>
         </div>
       </div>
-       <!-- Footer -->
-       <?php
+      <!-- Footer -->
+      <?php
       require_once('partials/_mainfooter.php');
       ?>
     </div>
